# 苍穹外卖-用户端接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/user/order/reminder/{id}": {
      "get": {
        "summary": "催单",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单id",
            "required": true,
            "example": "101",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R",
                  "$$ref": "#/definitions/R",
                  "required": [
                    "code"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/repetition/{id}": {
      "post": {
        "summary": "再来一单",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单id",
            "required": true,
            "example": "101",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R",
                  "$$ref": "#/definitions/R",
                  "required": [
                    "code"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/historyOrders": {
      "get": {
        "summary": "历史订单查询",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页面",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页记录数",
            "required": true,
            "example": "10",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "订单状态",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number"
                    },
                    "msg": {
                      "type": "null"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "records": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "number": {
                                "type": "string"
                              },
                              "status": {
                                "type": "number"
                              },
                              "userId": {
                                "type": "number"
                              },
                              "addressBookId": {
                                "type": "number"
                              },
                              "orderTime": {
                                "type": "string"
                              },
                              "checkoutTime": {
                                "type": "string"
                              },
                              "payMethod": {
                                "type": "number"
                              },
                              "payStatus": {
                                "type": "number"
                              },
                              "amount": {
                                "type": "number"
                              },
                              "remark": {
                                "type": "string"
                              },
                              "userName": {
                                "type": "null"
                              },
                              "phone": {
                                "type": "string"
                              },
                              "address": {
                                "type": "string"
                              },
                              "consignee": {
                                "type": "string"
                              },
                              "cancelReason": {
                                "type": "null"
                              },
                              "rejectionReason": {
                                "type": "null"
                              },
                              "cancelTime": {
                                "type": "null"
                              },
                              "estimatedDeliveryTime": {
                                "type": "string"
                              },
                              "deliveryStatus": {
                                "type": "number"
                              },
                              "deliveryTime": {
                                "type": "null"
                              },
                              "packAmount": {
                                "type": "number"
                              },
                              "tablewareNumber": {
                                "type": "number"
                              },
                              "tablewareStatus": {
                                "type": "number"
                              },
                              "orderDetailList": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "number"
                                    },
                                    "name": {
                                      "type": "string"
                                    },
                                    "orderId": {
                                      "type": "number"
                                    },
                                    "dishId": {
                                      "type": "number"
                                    },
                                    "setmealId": {
                                      "type": "null"
                                    },
                                    "dishFlavor": {
                                      "type": "null",
                                      "nullable": true
                                    },
                                    "number": {
                                      "type": "number"
                                    },
                                    "amount": {
                                      "type": "number"
                                    },
                                    "image": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "name",
                                    "orderId",
                                    "dishId",
                                    "setmealId",
                                    "dishFlavor",
                                    "number",
                                    "amount",
                                    "image"
                                  ]
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/cancel/{id}": {
      "put": {
        "summary": "取消订单",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单id",
            "required": true,
            "example": "101",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {}
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R",
                  "$$ref": "#/definitions/R",
                  "required": [
                    "code"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/orderDetail/{id}": {
      "get": {
        "summary": "查询订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "订单id",
            "required": true,
            "example": "101",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "address": {
                          "type": "string",
                          "refType": null
                        },
                        "addressBookId": {
                          "type": "integer",
                          "format": "int64",
                          "refType": null
                        },
                        "amount": {
                          "type": "number",
                          "refType": null
                        },
                        "cancelReason": {
                          "type": "string",
                          "refType": null
                        },
                        "cancelTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null
                        },
                        "checkoutTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null
                        },
                        "consignee": {
                          "type": "string",
                          "refType": null
                        },
                        "deliveryStatus": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "deliveryTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null
                        },
                        "estimatedDeliveryTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null
                        },
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "refType": null
                        },
                        "number": {
                          "type": "string",
                          "refType": null
                        },
                        "orderDetailList": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "amount": {
                                "type": "number",
                                "refType": null
                              },
                              "dishFlavor": {
                                "type": "string",
                                "refType": null
                              },
                              "dishId": {
                                "type": "integer",
                                "format": "int64",
                                "refType": null
                              },
                              "id": {
                                "type": "integer",
                                "format": "int64",
                                "refType": null
                              },
                              "image": {
                                "type": "string",
                                "refType": null
                              },
                              "name": {
                                "type": "string",
                                "refType": null
                              },
                              "number": {
                                "type": "integer",
                                "format": "int32",
                                "refType": null
                              },
                              "orderId": {
                                "type": "integer",
                                "format": "int64",
                                "refType": null
                              },
                              "setmealId": {
                                "type": "integer",
                                "format": "int64",
                                "refType": null
                              }
                            },
                            "title": "OrderDetail",
                            "$$ref": "#/definitions/OrderDetail"
                          },
                          "refType": "OrderDetail"
                        },
                        "orderTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null
                        },
                        "packAmount": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "payMethod": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "payStatus": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "phone": {
                          "type": "string",
                          "refType": null
                        },
                        "rejectionReason": {
                          "type": "string",
                          "refType": null
                        },
                        "remark": {
                          "type": "string",
                          "refType": null
                        },
                        "status": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "tablewareNumber": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "tablewareStatus": {
                          "type": "integer",
                          "format": "int32",
                          "refType": null
                        },
                        "userId": {
                          "type": "integer",
                          "format": "int64",
                          "refType": null
                        },
                        "userName": {
                          "type": "string",
                          "refType": null
                        }
                      },
                      "title": "OrderVO",
                      "$$ref": "#/definitions/OrderVO"
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R«OrderVO»",
                  "$$ref": "#/definitions/R«OrderVO»"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/submit": {
      "post": {
        "summary": "用户下单",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addressBookId": {
                    "type": "integer",
                    "format": "int64",
                    "refType": null,
                    "description": "地址簿id"
                  },
                  "amount": {
                    "type": "number",
                    "refType": null,
                    "description": "总金额"
                  },
                  "deliveryStatus": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "配送状态：  1立即送出  0选择具体时间"
                  },
                  "estimatedDeliveryTime": {
                    "type": "string",
                    "refType": null,
                    "description": "预计送达时间",
                    "example": "yyyy-MM-dd HH:mm:ss"
                  },
                  "packAmount": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "打包费"
                  },
                  "payMethod": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "付款方式"
                  },
                  "remark": {
                    "type": "string",
                    "refType": null,
                    "description": "备注"
                  },
                  "tablewareNumber": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "餐具数量"
                  },
                  "tablewareStatus": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "餐具数量状态  1按餐量提供  0选择具体数量"
                  }
                },
                "title": "OrdersSubmitDTO",
                "$$ref": "#/definitions/OrdersSubmitDTO",
                "required": [
                  "addressBookId",
                  "amount",
                  "deliveryStatus",
                  "estimatedDeliveryTime",
                  "packAmount",
                  "payMethod",
                  "remark",
                  "tablewareNumber",
                  "tablewareStatus"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64",
                          "refType": null,
                          "description": "订单id"
                        },
                        "orderAmount": {
                          "type": "number",
                          "refType": null,
                          "description": "订单金额"
                        },
                        "orderNumber": {
                          "type": "string",
                          "refType": null,
                          "description": "订单号"
                        },
                        "orderTime": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null,
                          "description": "下单时间"
                        }
                      },
                      "title": "OrderSubmitVO",
                      "$$ref": "#/definitions/OrderSubmitVO",
                      "required": [
                        "id",
                        "orderAmount",
                        "orderNumber",
                        "orderTime"
                      ]
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R«OrderSubmitVO»",
                  "$$ref": "#/definitions/R«OrderSubmitVO»",
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/user/order/payment": {
      "put": {
        "summary": "订单支付",
        "deprecated": false,
        "description": "",
        "tags": [
          "C端-订单接口"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderNumber": {
                    "type": "string",
                    "refType": null,
                    "description": "订单号"
                  },
                  "payMethod": {
                    "type": "integer",
                    "format": "int32",
                    "refType": null,
                    "description": "支付方式"
                  }
                },
                "title": "OrdersPaymentDTO",
                "$$ref": "#/definitions/OrdersPaymentDTO",
                "required": [
                  "orderNumber",
                  "payMethod"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "nonceStr": {
                          "type": "string",
                          "format": "date-time",
                          "refType": null,
                          "description": "随机字符串"
                        },
                        "paySign": {
                          "type": "string",
                          "description": "签名"
                        },
                        "timeStamp": {
                          "type": "string",
                          "description": "时间戳"
                        },
                        "signType": {
                          "type": "string",
                          "description": "签名算法"
                        },
                        "packageStr": {
                          "type": "string",
                          "description": "统一下单接口返回的 prepay_id 参数值"
                        }
                      },
                      "title": "OrderPaymentVO",
                      "$$ref": "#/definitions/OrderPaymentVO",
                      "required": [
                        "nonceStr",
                        "packageStr",
                        "signType",
                        "timeStamp",
                        "paySign"
                      ]
                    },
                    "msg": {
                      "type": "string"
                    }
                  },
                  "title": "R«OrderPaymentVO»",
                  "$$ref": "#/definitions/R«OrderPaymentVO»",
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}